+++ baybe/strategies/__init__.py
@@ -1,10 +1,15 @@
 """BayBE strategies."""
 
-from baybe.strategies.composite import SequentialStrategy, TwoPhaseStrategy
+from baybe.strategies.composite import (
+    SequentialStrategy,
+    StreamingSequentialStrategy,
+    TwoPhaseStrategy,
+)
 from baybe.strategies.deprecation import Strategy
 
 __all__ = [
     "SequentialStrategy",
+    "StreamingSequentialStrategy",
     "TwoPhaseStrategy",
     "Strategy",
 ]