+++ baybe/strategies/composite.py
@@ -1,15 +1,18 @@
 """Strategies that switch recommenders depending on the experimentation progress."""
 
-from typing import Iterable, Literal, Optional
+from typing import Iterable, Iterator, List, Literal, Optional
 
 import pandas as pd
 from attrs import define, field
-from attrs.validators import in_
+from attrs.validators import deep_iterable, in_, instance_of
 
+from baybe.exceptions import NoRecommendersLeftError
 from baybe.recommenders import RandomRecommender, SequentialGreedyRecommender
 from baybe.recommenders.base import Recommender
 from baybe.searchspace import SearchSpace
 from baybe.strategies.base import Strategy
+from baybe.utils import serialization_unsupported_hook
+from baybe.utils.serialization import converter
 
 
 @define(kw_only=True)
@@ -113,15 +116,29 @@ class SequentialStrategy(Strategy):
     A new recommender is taken from the sequence after each recommended batch until
     all recommenders are exhausted.
 
+    Note:
+        The provided sequence of recommenders will be internally pre-collected into a
+        list. If this is not acceptable, consider using
+        :class:`baybe.strategies.composite.StreamingSequentialStrategy` instead.
+
     Args:
-        recommenders: An iterable providing the recommenders to be used.
+        recommenders: A finite-length sequence of recommenders to be used.
+            (For infinite-length iterables, see
+            :class:`baybe.strategies.composite.StreamingSequentialStrategy`)
 
     Raises:
-        StopIteration: If more (batch) recommendations are requested than there are
-            recommenders available.
+        NoRecommendersLeftError: If more (batch) recommendations are requested than
+            there are recommenders available.
     """
 
-    recommenders: Iterable[Recommender] = field()
+    # Exposed
+    recommenders: List[Recommender] = field(
+        converter=list, validator=deep_iterable(instance_of(Recommender))
+    )
+
+    # Private
+    # TODO: See :class:`baybe.strategies.composite.TwoPhaseStrategy`.
+    _step: int = field(default=0, alias="_step")
 
     def select_recommender(  # noqa: D102
         self,
@@ -132,4 +149,68 @@ class SequentialStrategy(Strategy):
     ) -> Recommender:
         # See base class.
 
-        return next(self.recommenders)
+        try:
+            recommender = self.recommenders[self._step]
+        except IndexError as ex:
+            raise NoRecommendersLeftError(
+                f"The strategy has been queried {self._step+1} time(s) but the "
+                f"provided sequence of recommenders contains only "
+                f"{self._step} element(s)."
+            ) from ex
+        self._step += 1
+        return recommender
+
+
+@define(kw_only=True)
+class StreamingSequentialStrategy(Strategy):
+    """A strategy that switches between recommenders from an iterable.
+
+    Similar to :class:`baybe.strategies.composite.SequentialStrategy` but without
+    explicit list conversion. Consequently, it supports arbitrary iterables, possibly
+    of infinite length. The downside is that serialization is not supported.
+
+    Args:
+        recommenders: An iterable providing the recommenders to be used.
+
+    Raises:
+        StopIteration: If more (batch) recommendations are requested than there are
+            recommenders available.
+    """
+
+    # Exposed
+    recommenders: Iterable[Recommender] = field()
+
+    # Private
+    _iterator: Iterator = field(init=False)
+    _step: int = field(init=False, default=0)
+
+    @_iterator.default
+    def default_iterator(self):
+        """Initialize the recommender iterator."""
+        return iter(self.recommenders)
+
+    def select_recommender(  # noqa: D102
+        self,
+        searchspace: SearchSpace,
+        batch_quantity: int = 1,
+        train_x: Optional[pd.DataFrame] = None,
+        train_y: Optional[pd.DataFrame] = None,
+    ) -> Recommender:
+        # See base class.
+
+        try:
+            recommender = next(self._iterator)
+        except StopIteration as ex:
+            raise NoRecommendersLeftError(
+                f"The strategy has been queried {self._step+1} time(s) but the "
+                f"provided sequence of recommenders contains only "
+                f"{self._step} element(s)."
+            ) from ex
+        self._step += 1
+        return recommender
+
+
+# Prevent breaking serialization of the iterable
+converter.register_unstructure_hook(
+    StreamingSequentialStrategy, serialization_unsupported_hook
+)